Move Rateconverter and pvtIdx to the wellInterface
